requirements.txt
torch==2.5.1
torchvision==0.20.1
pytorch-lightning==2.5.2
# warmup_scheduler==0.3
huggingface_hub==0.27.1
datasets==3.1.0
einops==0.6.1
hydra-core==1.3.2
omegaconf==2.3.0
loguru==0.7.3
matplotlib==3.9.3
wandb==0.18.0
lpips==0.1.4
torchmetrics[image]==1.5.1
tqdm
